requirements.txt
https://github.com/kensoi/vkbotkit/tarball/dev
python-dotenv